+++ tests/integrations/examples/test_driftbif_simulation.py
@@ -5,12 +5,17 @@
 import numpy as np
 import unittest
 import pandas as pd
+import dask.dataframe as dd
 
 from tsfresh.examples.driftbif_simulation import velocity, load_driftbif, sample_tau
-from tsfresh import extract_relevant_features
+from tsfresh import extract_relevant_features, extract_features
+from tsfresh.feature_extraction import MinimalFCParameters
 
 
 class DriftBifSimlationTestCase(unittest.TestCase):
+    def setUp(self):
+        np.random.seed(42)
+
     def test_intrinsic_velocity_at_default_bifurcation_point(self):
         """
         The intrinsic velocity of a dissipative soliton at the Drift bifurcation point is zero.
@@ -65,7 +70,7 @@ class DriftBifSimlationTestCase(unittest.TestCase):
                          'The returned vector should reflect the dimension of the initial condition.')
 
     def test_relevant_feature_extraction(self):
-        df, y = load_driftbif(100, 10, classification=False, seed=42)
+        df, y = load_driftbif(100, 10, classification=False)
 
         df['id'] = df['id'].astype('str')
         y.index = y.index.astype('str')
@@ -78,6 +83,9 @@ class DriftBifSimlationTestCase(unittest.TestCase):
 
 
 class SampleTauTestCase(unittest.TestCase):
+    def setUp(self):
+        np.random.seed(42)
+
     def test_range(self):
         tau = sample_tau(100)
         self.assertGreaterEqual(min(tau), 2.87)
@@ -92,6 +100,9 @@ class SampleTauTestCase(unittest.TestCase):
 
 
 class LoadDriftBifTestCase(unittest.TestCase):
+    def setUp(self):
+        np.random.seed(42)
+
     def test_classification_labels(self):
         X, y = load_driftbif(10, 100)
         self.assertEqual(set(y), {0, 1})