+++ CHANGES.rst
@@ -7,6 +7,14 @@ tsfresh uses `Semantic Versioning <http://semver.org/>`_
 Unreleased
 ==========
 
+- Bugfixes:
+    - Do not send coverage of notebook tests to codecov #759
+    - Fix typos in notebook (#757)
+    - Fix output format of `make_forecasting_frame` (#758)
+
+Version 0.17.0
+==============
+
 We changed the default branch from "master" to "main".
 
 - Breaking Change
@@ -23,10 +31,17 @@ We changed the default branch from "master" to "main".
     - Add a test for the dask bindings (#719)
     - Refactor input data iteration to need less memory (#707)
     - Added benchmark tests (#710)
+    - Make dask a possible input format (#736)
 - Bugfixes:
     - Fixed a bug in the selection, that caused all regression tasks with un-ordered index to be wrong (#715)
     - Fixed readthedocs (#695, #696)
     - Fix spark and dask after #705 and for non-id named id columns (#712)
+    - Fix in the forecasting notebook (#729)
+    - Let tsfresh choose the value column if possible (#722)
+    - Move from coveralls github action to codecov (#734)
+    - Improve speed of data processing (#735)
+    - Fix for newer, more strict pandas versions (#737)
+    - Fix documentation for feature calculators (#743)
 
 Version 0.16.0
 ==============